Minor clean-ups/formatting
