+++ mobile/ios-app/Sources/Model/Objective.swift
@@ -1,13 +1,13 @@
 import Foundation
 
-struct Objective {
+struct Objective: Identifiable, Decodable {
     let id: Int
     let title: String
     let measurementType: MeasurementType
 }
 
 extension Objective {
-    enum MeasurementType {
+    enum MeasurementType: Decodable {
         case counter
         case duration
         case latency
