--- mobile/ios-app/Sources/UI/SessionView/CounterCard.swift
@@ -1,16 +1,0 @@
-import SwiftUI
-
-struct CounterCard: View {
-    @StateObject var counter: Counter
-    let color: Color
-    
-    var body: some View {
-        ObjectiveCard(objective: counter.objective, color: color, label: {
-            Text("\(counter.count())")
-                .font(.largeTitle)
-                .fontWeight(.bold)
-                .foregroundColor(.white)
-                .layoutPriority(100)
-        }, action: counter.increment)
-    }
-}
